Move ext file
