+++ common/design/build.gradle.kts
@@ -21,6 +21,7 @@
  */
 
 import dependencies.Dependencies
+import dependencies.Dependencies.AndroidX
 import dependencies.Dependencies.Compose
 import dependencies.Versions
 
@@ -53,6 +54,7 @@ android {
 
 dependencies {
     implementation(Dependencies.MATERIAL)
+    implementation(AndroidX.CORE)
 
     implementation(Compose.ui)
     implementation(Compose.material)
